backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
pymupdf==1.25.1
anthropic==0.42.0
chromadb==0.5.23
apscheduler==3.10.4
python-multipart==0.0.20
python-docx==1.1.2
python-pptx==1.0.2
olefile==0.47
